Use pipe to prevent automatically handle write pushback
When writing to a stream (e.g. stdout or a file stream) we normally have
to check the return value of each write to know wether or not we should
continue writing or wait until the buffer is free again.

When using a pipe we do not have to do that.

We now rewrite our markdown generator function to act as a generator
that yield chunks of markdown. This can be wrapped into a Readable that
can then easily be piped into a write stream.

Node should then automatically handle to only generate new data when the
write stream is actually ready to receive data.